#457 Move `JpaRepository` to dedicated JPA package
