+++ hartshorn-persistence/src/main/java/org/dockbox/hartshorn/persistence/service/JpaRepositoryFactory.java
@@ -20,7 +20,7 @@ package org.dockbox.hartshorn.persistence.service;
 import org.dockbox.hartshorn.core.annotations.Factory;
 import org.dockbox.hartshorn.core.annotations.inject.Enable;
 import org.dockbox.hartshorn.core.annotations.service.Service;
-import org.dockbox.hartshorn.persistence.JpaRepository;
+import org.dockbox.hartshorn.persistence.jpa.JpaRepository;
 
 @Service
 public interface JpaRepositoryFactory {
